Moves Pom class to common and adds source
